moved files from xsens to sonar
