+++ app/src/androidTest/java/sensors_in_paradise/sonar/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package sensors_in_paradise.xsens;
+package sensors_in_paradise.sonar;
 
 import android.content.Context;
