+++ app/src/test/java/sensors_in_paradise/sonar/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package sensors_in_paradise.xsens;
+package sensors_in_paradise.sonar;
 
 import org.junit.Test;
